kh12. A couple of leaks plugged on relay restarting, a possible crash on xml reload or server exit. small changes for win32, the icecast service should be working better and I've updated the libogg to 1.1.4 for the build.
svn path=/icecast/branches/kh/icecast/; revision=16352

+++ src/util.c
@@ -47,14 +47,14 @@
 
 struct rate_calc_node
 {
-    uint64_t index;
-    long value;
+    int64_t index;
+    uint64_t value;
     struct rate_calc_node *next;
 };
 
 struct rate_calc
 {
-    uint64_t total;
+    int64_t total;
     struct rate_calc_node *current;
     unsigned int samples;
     unsigned int ssec;
@@ -755,7 +755,7 @@ void rate_add (struct rate_calc *calc, long value, uint64_t sid)
  */
 long rate_avg (struct rate_calc *calc)
 {
-    uint64_t range;
+    float range;
 
     if (calc == NULL || calc->blocks < 2)
         return 0;
